chore: split meta evolution controller, regenerate generated data
