+++ data/runtime_stats.yaml
@@ -1,9 +1,9 @@
 schema_version: 1
-last_generated_utc: '2026-06-20T13:57:10Z'
-generator_revision: df30dacfb
+last_generated_utc: '2026-06-20T18:57:15Z'
+generator_revision: 4e7b5df39
 stats:
   tests:
-    raw: 37013
+    raw: 37100
     rounded: 37000
     display: 37,000+
   providers_curated:
